Move OCR script to ml file
